requirements.txt
async-timeout==4.0.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
spotipy==2.24.0
urllib3==2.2.2
Werkzeug==3.0.3
